# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Aaron Blog API 文檔'
  description: 'Aaron 個人博客系統的 API 文檔，提供文章、分類、標籤等功能的 RESTful API 介面。'
  contact:
    name: 'Aaron Blog 技術支援'
    email: admin@aaronblog.com
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8080'
    description: 'Aaron Blog API 伺服器'
paths:
...
  /api/admin/tags:
    post:
      tags:
        - Admin
      summary: 創建新標籤
      description: 創建新的文章標籤，需要管理員權限
      operationId: createTag
      requestBody:
        description: 標籤資料
        required: true
        content:
          application/json:
            schema:
              required:
                - name
                - slug
              properties:
                name:
                  description: 標籤名稱（唯一）
                  type: string
                  maxLength: 255
                  example: PHP
                slug:
                  description: '標籤 slug（唯一，只能包含字母、數字、破折號和底線）'
                  type: string
                  maxLength: 255
                  example: php
              type: object
      responses:
        201:
          description: 標籤創建成功
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: success
                  code:
                    type: integer
                    example: 201
                  message:
                    type: string
                    example: 操作成功
                  data:
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: PHP
                      slug:
                        type: string
                        example: php
                      articles_count:
                        type: integer
                        example: 0
                      created_at:
                        type: string
                        format: date-time
                        example: '2024-03-20T12:00:00.000000Z'
                    type: object
                type: object
        401:
          description: 未授權
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: error
                  code:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: Unauthenticated
                type: object
        422:
          description: 驗證失敗
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: error
                  code:
                    type: integer
                    example: 422
                  message:
                    type: string
                    example: 'The given data was invalid'
                  errors:
                    properties:
                      name:
                        type: array
                        items:
                          type: string
                          example: 標籤名稱已被使用
                      slug:
                        type: array
                        items:
                          type: string
                          example: 'slug 已被使用'
                    type: object
                type: object
      security:
        -
          sessionAuth: []
          csrfToken: []
  '/api/admin/tags/{id}':
    put:
      tags:
        - Admin
      summary: 更新標籤
      description: 更新指定的標籤資訊，需要管理員權限
      operationId: updateTag
      parameters:
        -
          name: id
          in: path
          description: '標籤 ID'
          required: true
          schema:
            type: integer
            minimum: 1
            example: 1
      requestBody:
        description: 更新的標籤資料（部分更新）
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  description: 標籤名稱（可選）
                  type: string
                  maxLength: 255
                  example: 'PHP Framework'
                slug:
                  description: '標籤 slug（可選）'
                  type: string
                  maxLength: 255
                  example: php-framework
              type: object
      responses:
        200:
          description: 標籤更新成功
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: success
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: 操作成功
                  data:
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: 'PHP Framework'
                      slug:
                        type: string
                        example: php-framework
                      articles_count:
                        type: integer
                        example: 5
                      created_at:
                        type: string
                        format: date-time
                        example: '2024-03-20T12:00:00.000000Z'
                      updated_at:
                        type: string
                        format: date-time
                        example: '2024-03-20T14:30:00.000000Z'
                    type: object
                type: object
        401:
          description: 未授權
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: error
                  code:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: Unauthenticated
                type: object
        404:
          description: 標籤不存在
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: error
                  code:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: 'Resource not found'
                type: object
        422:
          description: 驗證失敗
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: error
                  code:
                    type: integer
                    example: 422
                  message:
                    type: string
                    example: 'The given data was invalid'
                type: object
      security:
        -
          sessionAuth: []
          csrfToken: []
    delete:
      tags:
        - Admin
      summary: 刪除標籤
      description: 刪除指定的標籤，需要管理員權限
      operationId: deleteTag
      parameters:
        -
          name: id
          in: path
          description: '標籤 ID'
          required: true
          schema:
            type: integer
            minimum: 1
            example: 1
      responses:
        200:
          description: 標籤刪除成功
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: success
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: 操作成功
                  data:
                    example: null
                type: object
        401:
          description: 未授權
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: error
                  code:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: Unauthenticated
                type: object
        404:
          description: 標籤不存在
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: error
                  code:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: 'Resource not found'
                type: object
        409:
          description: 無法刪除（標籤仍被文章使用）
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: error
                  code:
                    type: integer
                    example: 409
                  message:
                    type: string
                    example: 'Cannot delete tag with associated articles'
                type: object
      security:
        -
          sessionAuth: []
          csrfToken: []
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      description: '使用 Laravel Session Cookie 進行身份驗證'
      name: laravel_session
      in: cookie
    csrfToken:
      type: apiKey
      description: 'CSRF 防護 Token，用於 POST/PUT/DELETE 等寫入操作'
      name: X-XSRF-TOKEN
      in: header
...